Added validation to screen resizing
While trying to keep a certain aspect ratio the console window was able
to grow above the max size, it can't now. I also renamed the Cons_Size
class because it is related more to Board_Size. Also added the ability
to save the board while running the game, and incidentally found and
fixed a bug where a population would not load from a file if it was the
same size as the board.